requirements.txt
```txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
sentence-transformers==2.2.2
openai-whisper==20231117
faiss-cpu==1.7.4
Pillow==10.1.0
opencv-python==4.8.1.78
librosa==0.10.1
streamlit==1.28.0
numpy==1.24.3
pandas==2.1.3
tqdm==4.66.1
```